Typecheck fix. NEXT: continue adding new indexes
