+++ webpack/resources/util.ts
@@ -1,5 +1,5 @@
 import { ResourceName } from "farmbot";
-import { joinKindAndId } from "./reducer";
+import { joinKindAndId } from "./reducer_support";
 import { Dictionary } from "farmbot/dist";
 import { betterCompact } from "../util";
 import * as _ from "lodash";
