+++ webpack/sequences/step_tiles/__tests__/tile_execute_test.tsx
@@ -8,7 +8,7 @@ import {
 import { mount } from "enzyme";
 import { fakeSequence } from "../../../__test_support__/fake_state/resources";
 import { Execute, Point, Identifier, Coordinate, Tool } from "farmbot";
-import { emptyState } from "../../../resources/reducer";
+import { emptyState } from "../../../resources/reducer_support";
 import { Actions } from "../../../constants";
 
 function fakeProps(): ExecBlockParams {
