+++ webpack/sequences/step_tiles/__tests__/tile_find_home_test.tsx
@@ -3,7 +3,7 @@ import { TileFindHome, FindHomeParams } from "../tile_find_home";
 import { mount } from "enzyme";
 import { fakeSequence } from "../../../__test_support__/fake_state/resources";
 import { FindHome } from "farmbot/dist";
-import { emptyState } from "../../../resources/reducer";
+import { emptyState } from "../../../resources/reducer_support";
 import {
   fakeHardwareFlags
 } from "../../../__test_support__/sequence_hardware_settings";
