+++ webpack/sequences/step_tiles/__tests__/tile_if_test.tsx
@@ -3,7 +3,7 @@ import { TileIf } from "../tile_if";
 import { mount } from "enzyme";
 import { fakeSequence } from "../../../__test_support__/fake_state/resources";
 import { If } from "farmbot/dist";
-import { emptyState } from "../../../resources/reducer";
+import { emptyState } from "../../../resources/reducer_support";
 
 describe("<TileIf/>", () => {
   function bootstrapTest() {
