+++ webpack/sequences/step_tiles/__tests__/tile_read_pin_test.tsx
@@ -3,7 +3,7 @@ import { TileReadPin } from "../tile_read_pin";
 import { mount } from "enzyme";
 import { fakeSequence } from "../../../__test_support__/fake_state/resources";
 import { ReadPin } from "farmbot/dist";
-import { emptyState } from "../../../resources/reducer";
+import { emptyState } from "../../../resources/reducer_support";
 
 describe("<TileReadPin/>", () => {
   function bootstrapTest() {
