+++ webpack/sequences/step_tiles/__tests__/tile_send_message_test.tsx
@@ -3,7 +3,7 @@ import { TileSendMessage, RefactoredSendMessage } from "../tile_send_message";
 import { mount } from "enzyme";
 import { fakeSequence } from "../../../__test_support__/fake_state/resources";
 import { SendMessage, Channel } from "farmbot/dist";
-import { emptyState } from "../../../resources/reducer";
+import { emptyState } from "../../../resources/reducer_support";
 import { channel } from "../tile_send_message_support";
 
 describe("<TileSendMessage/>", () => {
