+++ webpack/sequences/step_tiles/tile_if/__tests__/then_test.tsx
@@ -3,7 +3,7 @@ import { Then } from "../then";
 import { mount } from "enzyme";
 import { fakeSequence } from "../../../../__test_support__/fake_state/resources";
 import { If } from "farmbot/dist";
-import { emptyState } from "../../../../resources/reducer";
+import { emptyState } from "../../../../resources/reducer_support";
 import { IfParams } from "../index";
 
 describe("<Then/>", () => {
