+++ webpack/sequences/step_tiles/tile_move_absolute/generate_list.ts
@@ -11,7 +11,7 @@ import { Vector3 } from "farmbot/dist";
 import { TOOL } from "./interfaces";
 import * as _ from "lodash";
 import { t } from "i18next";
-import { joinKindAndId } from "../../../resources/reducer";
+import { joinKindAndId } from "../../../resources/reducer_support";
 import { capitalize } from "lodash";
 
 export function activeTools(resources: ResourceIndex) {
